Refactored ContentTabWidget into it's own file.
