+++ lib/routes/pages/games.dart
@@ -1,4 +1,5 @@
 import 'package:black_tortoise/routes/home.dart';
+import 'package:black_tortoise/widgets/tab_contents.dart';
 import 'package:black_tortoise/widgets/tabs.dart';
 import 'package:flutter/material.dart';
 
@@ -22,54 +23,6 @@ class GamesPage extends StatelessWidget {
             ContentTabWidget(child: TabContents()),
           ],
         ),
-      ),
-    );
-  }
-}
-
-class TabContents extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 100,
-      decoration: BoxDecoration(borderRadius: BorderRadius.circular(15)),
-      child: Row(
-        children: [
-          Flexible(
-            flex: 2,
-            child: Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(15),
-                color: Colors.red,
-              ),
-            ),
-          ),
-          Flexible(
-            flex: 5,
-            child: Column(
-              children: [
-                Flexible(
-                  flex: 2,
-                  child: Container(
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(15),
-                      color: Colors.blue,
-                    ),
-                  ),
-                ),
-                Flexible(
-                  flex: 5,
-                  child: Container(
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(15),
-                      color: Colors.amber,
-                    ),
-                  ),
-                )
-              ],
-            ),
-          )
-        ],
       ),
     );
   }
